added configuration file refactored the flows
