+++ src/main/resources/log4j.properties
@@ -1,0 +1,13 @@
+es#=Root logger option
+log4j.rootLogger=INFO, stdout
+# Direct log messages to stdout
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.Target=System.out
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=%5p | %d  | %c{1}:%L | %m%n
+log4j.logger.Remoting=ERROR
+log4j.logger.akka.event.slf4j=DEBUG
+log4j.logger.com.anjuke.dm=${dm.logging.level}
+log4j.logger.org.eclipse.jetty=WARN
+log4j.logger.org.eclipse.jetty.util.component.AbstractLifeCycle=ERROR
+log4j.logger.com.datastax.driver.core.QueryLogger.SLOW=ERROR
