+++ src/main/resources/storage.conf
@@ -1,0 +1,6 @@
+
+cassandra {
+  host = "127.0.0.1"
+  port = 9042
+}
+
